working on role base authentication on different controllers

+++ app/Http/Controllers/ClientController.php
@@ -5,6 +5,7 @@ namespace People\Http\Controllers;
 use Illuminate\Http\Request;
 use People\Models\Client;
 use People\Services\Interfaces\IClientService;
+use People\Services\Interfaces\IUserAuthenticationService;
 
 class ClientController extends Controller
 {
@@ -15,23 +16,31 @@ class ClientController extends Controller
      */
     public $ClientService;
 
-    public function __construct(IClientService $clientService)
+    public function __construct(IClientService $clientService, IUserAuthenticationService $userAuthenticationService)
     {
-
-        $this->ClientService = $clientService;
+        $this->middleware('auth');
+        $this->ClientService             = $clientService;
+        $this->UserAuthenticationService = $userAuthenticationService;
     }
     public function showClientForm($companyId)
     {
+        $isAdmin = $this->UserAuthenticationService->isAdmin();
+        if ($isAdmin) {
+            $clients = $this->ClientService->getAllClients();
 
+            return view('clients.index',
+                [
+                    'clients'   => $clients,
+                    'companyId' => $companyId,
 
-        $clients = $this->ClientService->getAllClients();
-
-        return view('clients.index',
-            [
-                'clients' => $clients,
-                'companyId' => $companyId,
-
-            ]);
+                ]);
+        } else {
+            return view('notAuthorize',
+                [
+                    'message' => 'You are Not Authorize to view this Page !!',
+                ]
+            );
+        }
     }
     public function index(Request $request)
     {
@@ -40,7 +49,7 @@ class ClientController extends Controller
 
         return view('clients.index',
             [
-                'clients' => $clients,
+                'clients'   => $clients,
                 'companyId' => $request->companyId,
 
             ]);
@@ -64,12 +73,21 @@ class ClientController extends Controller
      */
     public function store(Request $request)
     {
-        $this->validate($request, array(
-            'name'  => 'required|max:255',
-        ));
-        $clientId = $this->ClientService->createClient($request);
+        $isAdmin = $this->UserAuthenticationService->isAdmin();
+        if ($isAdmin) {
+            $this->validate($request, array(
+                'name' => 'required|max:255',
+            ));
+            $clientId = $this->ClientService->createClient($request);
 
-        return redirect('/clients/' . $clientId);
+            return redirect('/clients/' . $clientId);
+        } else {
+            return view('notAuthorize',
+                [
+                    'message' => 'You are Not Authorize to view this Page !!',
+                ]
+            );
+        }
     }
 
     /**
@@ -84,9 +102,9 @@ class ClientController extends Controller
         $clientProjects = $this->ClientService->getClientProjects($client);
 
         return view('clients/showClient',
-            ['client' => $client,
+            ['client'        => $client,
                 'clientProjects' => $clientProjects,
-                'companyId' => $request->companyId,
+                'companyId'      => $request->companyId,
             ]);
     }
 
@@ -98,13 +116,22 @@ class ClientController extends Controller
      */
     public function edit(Client $client, Request $request)
     {
+        $isAdmin = $this->UserAuthenticationService->isAdmin();
+        if ($isAdmin) {
 
-        return view('clients/clientEditForm',
-            [
-                'client' => $client,
-                'companyId' => $request->companyId,
+            return view('clients/clientEditForm',
+                [
+                    'client'    => $client,
+                    'companyId' => $request->companyId,
 
-            ]);
+                ]);
+        } else {
+            return view('notAuthorize',
+                [
+                    'message' => 'You are Not Authorize to view this Page !!',
+                ]
+            );
+        }
     }
 
     /**
@@ -116,12 +143,22 @@ class ClientController extends Controller
      */
     public function update(Request $request, Client $client)
     {
-         $this->validate($request, array(
-            'name'  => 'required|max:255',
-        ));
-        $this->ClientService->updateClient($request, $client);
+        $isAdmin = $this->UserAuthenticationService->isAdmin();
+        if ($isAdmin) {
 
-        return redirect('/clients/' . $client->id);
+            $this->validate($request, array(
+                'name' => 'required|max:255',
+            ));
+            $this->ClientService->updateClient($request, $client);
+
+            return redirect('/clients/' . $client->id);
+        } else {
+            return view('notAuthorize',
+                [
+                    'message' => 'You are Not Authorize to view this Page !!',
+                ]
+            );
+        }
     }
     /**
      * Remove the specified resource from storage.
@@ -132,10 +169,18 @@ class ClientController extends Controller
 
     public function destroy(Client $client)
     {
+        $isAdmin = $this->UserAuthenticationService->isAdmin();
+        if ($isAdmin) {
 
-        $this->ClientService->deleteClient($client);
+            $this->ClientService->deleteClient($client);
 
-        return redirect('/companies/' . $client->company_id);
-
+            return redirect('/companies/' . $client->company_id);
+        } else {
+            return view('notAuthorize',
+                [
+                    'message' => 'You are Not Authorize to view this Page !!',
+                ]
+            );
+        }
     }
 }
